move plot_vs_posterior_plot
