+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
           - --exclude=docs/
           - --exclude=scripts/
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.9.1
+    rev: v0.9.2
     hooks:
       - id: ruff
         types_or: [python, pyi, jupyter]
